refactor: 날짜 출력 함수 별도 파일로 분리
