+++ src/pages/Diaries/New/Write.tsx
@@ -1,5 +1,6 @@
 import styled from '@emotion/styled';
 import { useMemo } from 'react';
+import formatKRDate from '../constants/formatKRDate';
 
 const DateText = styled.p`
   text-align: center;
@@ -85,13 +86,6 @@ function DiariesNewWrite() {
       <NextButton>다음</NextButton>
     </>
   );
-}
-
-function formatKRDate(d: Date) {
-  const y = d.getFullYear();
-  const m = `${d.getMonth() + 1}`.padStart(2, '0');
-  const day = `${d.getDate()}`.padStart(2, '0');
-  return `${y}년 ${m}월 ${day}일`;
 }
 
 export default DiariesNewWrite;
