+++ src/pages/Diaries/constants/formatKRDate.tsx
@@ -1,0 +1,8 @@
+function formatKRDate(d: Date) {
+  const y = d.getFullYear();
+  const m = `${d.getMonth() + 1}`.padStart(2, '0');
+  const day = `${d.getDate()}`.padStart(2, '0');
+  return `${y}년 ${m}월 ${day}일`;
+}
+
+export default formatKRDate;
